% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr.R
\name{mlr}
\alias{mlr}
\title{mlr}
\usage{
mlr(formula, data, na.action = "na.omit")
}
\arguments{
\item{formula}{the formula that linear regression model should fit}

\item{data}{the dataframe where the input data comes from}

\item{na.action}{a function which indicates what should happen when the data contain NAs. Default is na.omit.}
}
\value{
a list that contains the following values: coefficients, residuals, rank, fitted.values, df.residual, call, terms, model
}
\description{
Fits the multiple linear regression model
}
\examples{
mlr(mpg ~ cyl + wt, mtcars)
mlr(mpg ~ cyl + wt + cyl * wt, mtcars)
mlr(mpg ~ cyl + wt, mtcars, na.action = "na.exclude")

}
