% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_datetime.R
\name{mlr_business_days}
\alias{mlr_business_days}
\title{Return a count for the number of business days between two dates.}
\usage{
mlr_business_days(start, end)
}
\arguments{
\item{start}{: The start date for counting business days. Must be type
character}

\item{end}{: The end date for counting business days. Must be type
character}
}
\description{
This function uses the package `timeDate` to return a count of the
number of business days between two dates.
}
