% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_chartdata.R
\name{mlr_chart_data}
\alias{mlr_chart_data}
\title{Converts key:value pairs into ChartData specification written by developers at BigDatr}
\usage{
mlr_chart_data(dataframe, labels = ...)
}
\arguments{
\item{dataframe}{the data to be used for rows. Must be a column for each key:value pair
with column name as the key and row as the value.}

\item{labels}{a `c()` must be passed for each label you are assigning in columns.}
}
\value{
Outputs a JSON structure with Key:Value pairs in rows and columns. This JSON
structure can then be written to a file, s3, database etc. This function will not perform
this final step.
}
\description{
Converts key:value pairs into ChartData specification written by developers at BigDatr
}
\examples{
mlr_chart_data(no_of_users, c("Date", "Users"))
}
