% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{StandardError}
\alias{StandardError}
\title{Gets the Standard Error from a coefficient variance matrix}
\usage{
StandardError(x, n)
}
\arguments{
\item{x}{Numeric vector in R}

\item{n}{integer in R}
}
\description{
Gets the Standard Error from a coefficient variance matrix
}
