% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_mlr.R
\name{summary_mlr}
\alias{summary_mlr}
\title{summary_mlr}
\usage{
summary_mlr(model)
}
\arguments{
\item{model}{the model generated from the mlr function in this package}
}
\value{
a list that contains the following values:     terms , residuals, coefficients , sigma, df,  r.squared, adj.r.squared, fstatistic
}
\description{
Gives the summary result of mlr function
}
\examples{
summary_mlr(mlr(mpg ~ cyl + wt, mtcars))

}
